cmd/nri-kubernetes: move integration code to CLI package
